Added new Client options for connection pool config
The default connection pool size in Python requests is 10. This is
insufficient for high volume requests because the pool will be quickly
exhausted with an "HttpConnectionPool is full, discarding connection:
api.tempo-db.com" error. This allows the user to configure the poolsize
to meet their own requirements if the default is insufficient.